Allow negative LED index to be ignored
Some platforms don't have all LEDs. In order to indicate that an LED doesn't exist, the index can be set to -1. This change allows requests for that invalid index to be ignored.

+++ components/include/lcz_led.h
@@ -23,7 +23,7 @@ extern "C" {
 /******************************************************************************/
 /* Global Constants, Macros and Type Definitions                              */
 /******************************************************************************/
-typedef size_t led_index_t;
+typedef int led_index_t;
 
 /* The on and off times are in system ticks.
  * If the repeat count is 2 the pattern will be displayed 3
